Solution for Day 15 2015
This took a little work to get it to be generic enough for the
number of ingredients. The example only had 2 ingredients while
the actual input had 4. Also, didn't originally make sure that the
teaspoons of the ingredients equaled 100. I allowed cases where
the total was less than 100. That was really slow. Once I made the
fix, it was much faster.