Solution for Day 21 2015
Ended up getting part 1 pretty easily. Had problems with part 2
because I missed two parts of the acutal rules. The first rule
was that you HAD to have a weapon. The second was that I split the
rings into damage rings and defense rings and didn't take into
account that you could actually have 2 of those types, instead of
one of each. Fortunately, the data worked right for part 1 with
the bad logic. Fixed the weapon one and got closer to the answer,
but still wasn't right. Once I realized the ring one, it was
pretty simple getting the logic right. Then did a bunch of clean up.
Maybe did a little too much. :)