Solution for Day 22 2015
Finally got back to this one and got it done. I was kind of scared
of it, because it just seemed more complicated, especially with
having to deal with the logic of selecting the right spell. Finally
just sat down and worked it out on paper. The unit tests of the
examples definitely helped.