Solution for Day 9 2025
Part 2 was really hard. I started off by trying to use the ray
casting algorithm to determine if a point was in the polygon. That
worked for the sample input, but nothing I did could get it to
work for the real input. I went to HyperNeutrino's video on YouTube
(https://www.youtube.com/watch?v=toDrFDh7VNs) to get the solution.
It's a really nifty way to do it with coordinate compression and
a prefix sum matrix. So, thanks HyperNeutrino.