Solution for Day 12 2025
Advent of Code 2025 is done! This last puzzle was evil. It makes
you think you need to do some complex logic, but it doesn't. The
code doesn't work with the sample input, but it doesn't have to.
I left the sample input in the test case, even thought it fails.